UI 수정

+++ Assets/Data/CardData.cs
@@ -7,8 +7,8 @@ public class CardData : ScriptableObject
 {
     [SerializeField] Character character;
     public Character Char { get { return character; } }
-    [SerializeField] int speed;
-    public int Speed { get; set; }
+    [SerializeField] float speed;
+    public float Speed { get; set; }
 
     [SerializeField] int cost;
     public int Cost { get; set; }
